Add unit tests (#1)
