+++ requirements.test.txt
@@ -1,0 +1,2 @@
+pytest
+pytest-homeassistant-custom-component
