+++ salusfy/mock_ha_web_client.py
@@ -1,0 +1,7 @@
+
+class MockHaWebClient:
+    def __init__(self):
+        pass
+
+    def current_temperature(self):
+        return 15.9
