+++ tests/entity_registry.py
@@ -1,0 +1,13 @@
+class EntityRegistry:
+    def __init__(self):
+        self._entities = []
+    
+    def register(self, list):
+        self._entities.extend(list)
+    
+    @property
+    def entities(self):
+        return self._entities
+
+    def first(self):
+        return self._entities[0]
