+++ tests/mock_config.py
@@ -1,0 +1,7 @@
+USERNAME = "john@smith.com"
+PASSWORD = "12345"
+DEVICE_ID = "999999"
+ENTITY_ID = "sensor.everything_presence_one_temperature"
+ACCESS_TOKEN = "some-secret"
+SIMULATOR = True
+HOST = "192.168.0.99"
